requirements.txt
allure-pytest==2.5.0
allure-python-commons==2.5.0
apipkg==1.5
atomicwrites==1.1.5
attrs==18.1.0
blindspin==2.0.1
cached-property==1.4.3
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
crayons==0.1.2
docker==3.4.1
docker-pycreds==0.3.0
docopt==0.6.2
enum34==1.1.6
execnet==1.5.0
Faker==0.8.17
future==0.16.0
idna==2.6
jsonschema==2.6.0
more-itertools==4.2.0
pluggy==0.7.1
py==1.5.4
pypiwin32==220
pytest==3.6.4
pytest-cache==1.0
pytest-logger==0.3.0
python-dateutil==2.7.3
PyYAML==3.13
requests==2.18.4
selenium==2.53.1
six==1.11.0
StringGenerator==0.3.1
testcontainers==2.2
text-unidecode==1.2
texttable==0.9.1
urllib3==1.22
webdriver-manager==1.7
websocket-client==0.48.0
wrapt==1.10.11
